Refactor: TrendingPostAppender 분리
